+++ flutter_osm_interface/lib/src/osm_controller/osm_controller.dart
@@ -22,6 +22,7 @@ abstract class IBaseOSMController {
   Future<void> addMarker(
     GeoPoint p, {
     MarkerIcon? markerIcon,
+    double? angle,
   });
 
   ///remove marker from map of position